This commit contains Changes made to perfect the create A new entry function
I made an array of posts, that contains all the posts. Each Post is an object that has properties like the date, title, content , tag.
I was able to pass this array to the ejs file by enclosing it in array.

I used a forEACH loop to render all the pre-made posts in the array on the home page.
When a new post is created it's unshfited to the beginning of the array and hence is shown as the first post. With it's date saved.

The same index.js page is rendered as a response to the POST /createNewEntry.

For the Post Content, I changed it from an input to a textarea element, so we can see multiple lines as the user types.

I also created a  new ejs file called reading view, this is to show the full post when it's clicked.